Refactor NewCourseForm components for course creation
- Introduced BasicCourseInfo, DuplicateOptions, LanguageSelection, and LanguageVersionOptions components to modularize the NewCourseForm functionality.
- Implemented form handling for course details, duplication options, and language selection, enhancing user experience and maintainability.
- Removed the legacy NewCourseForm component to streamline the codebase and improve organization.

This change enhances the structure of the course creation process, making it more intuitive and easier to manage.
