+++ shared-module/packages/common/src/locales/en/main-frontend.json
@@ -134,6 +134,7 @@
   "confirm-reset-message": "Are you sure you want to reset exercises for the selected users?",
   "confirm-reset-title": "Confirm Reset",
   "confirm-your-password": "Confirm your password",
+  "copied-course-is-a-language-version": "Copied course is a language version",
   "could-not-find-course-course-instance-or-user-course-settings-for-enrollment": "Could not find course or course instance or user course settings for this enrollment.",
   "course": "Course",
   "course-created-successfully": "Course created successfully",
@@ -146,6 +147,7 @@
   "course-pages-for": "Course pages for {{course-name}}",
   "course-status-summary": "Course status summary",
   "course-version-selection": "Course version selection",
+  "course-where-to-copy-the-content": "Course where to copy the content",
   "courses": "Courses",
   "courses-completed-by-users-from": "Courses completed by users from",
   "create-an-acount": "Create an account",
@@ -555,7 +557,6 @@
   "new-course": "New course",
   "new-course-instance": "New course instance",
   "new-course-instance-dialog": "New Course Instance dialog",
-  "new-course-is-a-language-version": "New course is a language version",
   "new-definition": "New definition",
   "new-exam": "New Exam",
   "new-exam-dialog": "New Exam dialog",
@@ -667,7 +668,7 @@
   "research-consent-title": "Regarding research done on courses",
   "reset-my-own-progress-on-the-course": "Reset my own progress on the course",
   "reset-progress-for-all-students-on-the-course-draft": "Reset progress for all students on the course (works only on draft courses)",
-  "resulting-course-should-be-a-language-version-of-a-different-course": "The resulting course should be a language version of a different course than the one the material was copied from.",
+  "resulting-course-should-be-a-language-version-of-a-different-course": "The resulting course should be a language version of a different course than the one the material was copied from",
   "reveal": "Reveal",
   "role-admin": "Admin",
   "role-assistant": "Assistant",
@@ -817,7 +818,7 @@
   "suspected-cheater": "Suspected Cheaters",
   "suspected-student": "Suspected student",
   "swedish": "Swedish",
-  "target-course-id": "Target course ID",
+  "target-course": "Target course",
   "teacher-has-graded-this-manually": "Teacher has graded this manually",
   "teacher-in-charge-email": "Teacher in charge email",
   "teacher-in-charge-name": "Teacher in charge name",
